feat(run): проектный файл заводит человек, словарь теряет две операции
`run` — это то, для чего нужна платформа или база. Граница была с оговоркой:
`workspace.initialize` платформы не требует, а `source.create` создаёт
XML-каркас файлами, то есть пишет исходники — дом такой работы `apply`. Обе
сняты, и граница сходится без исключений.

`v8project.yaml` создаётся и правится руками — человеком в редакторе или
моделью её файловыми средствами. Инструмента записи проектного файла в
продукте не остаётся вовсе, и это осознанное состояние, а не пробел.

Ничего не ломается. `view {}` уже работает без файла: отвечает
`config.state: "autodetected"`, перечисляет найденные наборы и несёт в `setup`
рекомендуемое содержимое. Из `next` уходит подсказка-действие: ответ больше не
называет вызова, которым файл появится. Вопрос в `next` остаётся — `check {}`,
потому что вердикт о готовности живёт там, и его спрашивают, а не исполняют.

Словарь сходится с двенадцати операций до десяти, реализованных — с трёх до
двух. Это честнее прежнего счёта, где реализованной числилась операция,
которой не будет.

Модуль исполнителя переименован по новому предмету: он теперь отдаёт словарь и
отказывает по неизвестному имени.
