+++ arch/decisions/2026-09-09-project-config-is-handwritten.md
@@ -1,0 +1,40 @@
+---
+id: DEC.2026-09-09.PROJECT-CONFIG-IS-HANDWRITTEN
+status: active
+governs: product
+realized: crates/unica-coder/src/application/v13/tool_catalog.rs::v13_run_dictionary_has_twelve_directional_runtime_intents
+supersedes: []
+superseded-by: null
+establishes: [INV.WIRE.RUN-NEEDS-THE-PLATFORM]
+changes: [CTR.WIRE.TOOL-SURFACE]
+design: docs/plans/2026-09-04-run-operations-requirements.md
+---
+
+# Проектный файл заводит человек, а не инструмент
+
+**Решение.** Операции `workspace.initialize` и `source.create` снимаются со
+словаря `run`. `v8project.yaml` создаётся и правится руками — человеком в
+редакторе или моделью её файловыми средствами. Инструмента записи проектного
+файла в продукте не остаётся вовсе, и это осознанное состояние, а не пробел.
+Отдельный вход рассматривается не раньше 0.14 и тогда будет называться
+`unica.workspace`, а не операцией `run`.
+
+**Что чинит.** Граница `run` была с оговоркой. `run` — это то, для чего нужна
+платформа или база; `workspace.initialize` платформы не требует и оставался
+особым случаем. `source.create` создаёт XML-каркас файлами — это запись
+исходников, и её дом `apply`. После снятия обеих граница сходится без
+исключений: в словаре не остаётся ни одной операции, которая обходится без
+платформы или базы.
+
+**Чем держится.** `unica.view {}` уже работает без проектного файла: отвечает
+`config.state: "autodetected"`, перечисляет найденные наборы и несёт в `setup`
+рекомендуемое содержимое файла. Всё, что делала снимаемая операция, читатель
+ответа теперь делает сам, и ему для этого достаточно того, что уже в ответе.
+
+**Цена.** Из `next` у `unica.view {}` уходит подсказка-действие: ответ больше
+не называет вызова, которым файл появится. Вопрос в `next` остаётся —
+`unica.check {}`, потому что вердикт о готовности живёт там. Разница
+принципиальная: `check` спрашивают, а не исполняют. Словарь при этом сходится
+с двенадцати операций до десяти, а число реализованных — с трёх до двух: это
+честнее прежнего счёта, где реализованной числилась операция, которой не
+будет.
