+++ arch/index.md
@@ -111,6 +111,7 @@
 | `DEC.2026-09-08.SEARCH-PROVIDER-ROLES` | решение · product | active | да | Роль называет, чем искать, и без роли поиск остаётся своим | [decisions/2026-09-08-search-provider-roles.md](decisions/2026-09-08-search-provider-roles.md) |
 | `DEC.2026-09-08.TEMPLATE-CELL-CONTENT` | решение · product | active | да | Текст ячеек макета — адрес, а не признак | [decisions/2026-09-08-template-cell-content.md](decisions/2026-09-08-template-cell-content.md) |
 | `DEC.2026-09-08.V5-RUNTIME-HOOKS` | решение · product | active | да | Рантайм v5 наблюдаем через один объект крючков, а не через признак сборки | [decisions/2026-09-08-v5-runtime-hooks.md](decisions/2026-09-08-v5-runtime-hooks.md) |
+| `DEC.2026-09-09.PROJECT-CONFIG-IS-HANDWRITTEN` | решение · product | active | да | Проектный файл заводит человек, а не инструмент | [decisions/2026-09-09-project-config-is-handwritten.md](decisions/2026-09-09-project-config-is-handwritten.md) |
 | `INV.APP.ACTOR-AUTHENTICATED-SOURCE-CAPABILITIES` | инвариант · product | active |  | Binding определяет весь исполняемый профиль источника | [invariants/INV.APP.ACTOR-AUTHENTICATED-SOURCE-CAPABILITIES.md](invariants/INV.APP.ACTOR-AUTHENTICATED-SOURCE-CAPABILITIES.md) |
 | `INV.APP.ACTOR-AUTHENTICATED-SOURCE-IDENTITY` | инвариант · product | active |  | Актор сохраняет полный typed source profile | [invariants/INV.APP.ACTOR-AUTHENTICATED-SOURCE-IDENTITY.md](invariants/INV.APP.ACTOR-AUTHENTICATED-SOURCE-IDENTITY.md) |
 | `INV.APP.CODE-DEFINITION-READINESS` | инвариант · product | active |  | Definition не публикует ложный типизированный успех | [invariants/INV.APP.CODE-DEFINITION-READINESS.md](invariants/INV.APP.CODE-DEFINITION-READINESS.md) |
@@ -389,6 +390,7 @@
 | `INV.WIRE.PREVIEW-IS-MUTATION-ONLY` | инвариант · product | active |  | Предпросмотр принадлежит мутации | [invariants/INV.WIRE.PREVIEW-IS-MUTATION-ONLY.md](invariants/INV.WIRE.PREVIEW-IS-MUTATION-ONLY.md) |
 | `INV.WIRE.RESOLVE-EXACT-BRIDGE` | инвариант · product | active |  | Путь живёт в одном инструменте и отвечает точно | [invariants/INV.WIRE.RESOLVE-EXACT-BRIDGE.md](invariants/INV.WIRE.RESOLVE-EXACT-BRIDGE.md) |
 | `INV.WIRE.ROOT-FACTS-AND-VERDICT` | инвариант · product | active |  | Корень не смешивает факты с вердиктом | [invariants/INV.WIRE.ROOT-FACTS-AND-VERDICT.md](invariants/INV.WIRE.ROOT-FACTS-AND-VERDICT.md) |
+| `INV.WIRE.RUN-NEEDS-THE-PLATFORM` | инвариант · product | active |  | В словаре `run` нет операции, которая обходится без платформы | [invariants/INV.WIRE.RUN-NEEDS-THE-PLATFORM.md](invariants/INV.WIRE.RUN-NEEDS-THE-PLATFORM.md) |
 | `INV.WIRE.SDK-DEPENDENCY` | инвариант · process | active |  | Зависимость от SDK принадлежит только транспортному package | [invariants/INV.WIRE.SDK-DEPENDENCY.md](invariants/INV.WIRE.SDK-DEPENDENCY.md) |
 | `INV.WIRE.SDK-INITIALIZE` | инвариант · product | active |  | Legacy initialize публикует имя, версию сборки и согласованный протокол | [invariants/INV.WIRE.SDK-INITIALIZE.md](invariants/INV.WIRE.SDK-INITIALIZE.md) |
 | `INV.WIRE.SDK-MODULE-EXPORTS` | инвариант · product | active |  | Транспортный модуль сохраняет точный legacy Rust API | [invariants/INV.WIRE.SDK-MODULE-EXPORTS.md](invariants/INV.WIRE.SDK-MODULE-EXPORTS.md) |
