+++ crates/unica-coder/src/infrastructure/daemon/mod.rs
@@ -8,10 +8,10 @@ pub(crate) mod server;
 pub(crate) mod terminal_codec_v5;
 mod v13_infobase_exports;
 mod v13_read_modes;
+mod v13_run_dictionary;
 #[allow(dead_code)]
 mod v13_service;
 mod v13_workspace_bootstrap;
-mod v13_workspace_initialize;
 
 use identity::CoreIdentity;
 use std::path::Path;
