+++ crates/unica-coder/src/infrastructure/daemon/server.rs
@@ -302,10 +302,7 @@ impl V5CanonicalInvocationRuntime {
         {
             return Err(V5CanonicalPrepareError::Direct(Box::new(result)));
         }
-        if let Some(result) = super::v13_workspace_initialize::execute_workspace_initialize(
-            &request,
-            &response_deadline,
-        ) {
+        if let Some(result) = super::v13_run_dictionary::execute_run_dictionary(&request) {
             return Err(V5CanonicalPrepareError::Direct(Box::new(result)));
         }
         match super::v13_infobase_exports::prepare(&request) {
@@ -341,9 +338,7 @@ impl V5CanonicalInvocationRuntime {
             }
             WorkspaceAdmissionError::Invalid => {
                 if let Some(result) =
-                    super::v13_workspace_initialize::reject_unavailable_run_before_admission(
-                        &request,
-                    )
+                    super::v13_run_dictionary::reject_unavailable_run_before_admission(&request)
                 {
                     V5CanonicalPrepareError::Direct(Box::new(result))
                 } else {
@@ -3909,17 +3904,18 @@ struct ActorLogicalReadLease {"#,
             .and_then(|data| data.get("operations"))
             .and_then(serde_json::Value::as_array)
             .expect("run dictionary has operations");
-        assert_eq!(
+        // Проектный файл заводит человек, поэтому операции с таким именем в
+        // словаре нет вовсе — ни реализованной, ни объявленной.
+        assert!(
             operations
                 .iter()
-                .find(|operation| operation["op"] == "workspace.initialize")
-                .and_then(|operation| operation["implemented"].as_bool()),
-            Some(true)
+                .all(|operation| operation["op"] != "workspace.initialize"),
+            "{operations:?}"
         );
         assert!(
             operations.iter().all(|operation| !matches!(
                 operation["op"].as_str(),
-                Some("syntax.check" | "test.run" | "query.execute")
+                Some("syntax.check" | "test.run" | "query.execute" | "source.create")
             )),
             "v0.13 Run discovery must omit deferred check/test/query execution: {operations:?}"
         );
