+++ crates/unica-coder/src/infrastructure/daemon/v13_service.rs
@@ -1008,7 +1008,7 @@ impl CanonicalV13ReadService {
         let arguments = invocation.arguments();
         let catalog = catalog_for(SurfaceRelease::V13).expect("canonical catalog exists");
         let Some(op) = arguments.get("op") else {
-            return super::v13_workspace_initialize::run_dictionary_result();
+            return super::v13_run_dictionary::run_dictionary_result();
         };
         let Some(op) = op.as_str() else {
             return error_result(None, RefusalCode::BadValue, "run op must be a string");
